# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Testing api",
        "version": "1.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "User log In and return generated token",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "User logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/registration": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registration",
                "description": "User registration",
                "operationId": "registration",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegistrationRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegistrationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/favorites": {
            "get": {
                "tags": [
                    "Favorite"
                ],
                "summary": "Get list of favorite products",
                "description": "Returns list of favorite products",
                "operationId": "52c99046e948b195720aad5b63b772a6",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/favorites/{product}": {
            "post": {
                "tags": [
                    "Favorite"
                ],
                "summary": "Add product  favorite list",
                "description": "Add product  favorite list",
                "operationId": "a3e96926e8dd69bc22dd118f123ce937",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/favorites/{favorite}": {
            "delete": {
                "tags": [
                    "Favorite"
                ],
                "summary": "Remove product from favorite list",
                "description": "Remove product from favorite list",
                "operationId": "0bc677227adb44c276fbe1b1b8c91a81",
                "parameters": [
                    {
                        "name": "favorite",
                        "in": "path",
                        "description": "Favorite id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/products": {
            "get": {
                "tags": [
                    "Auth Product"
                ],
                "summary": "Get list of products",
                "description": "Returns list of products",
                "operationId": "790b4d045eb44933de2a95b669d3958f",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Auth Product"
                ],
                "summary": "Store Product",
                "description": "Store Product",
                "operationId": "store product",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/products/{product}": {
            "post": {
                "tags": [
                    "Auth Product"
                ],
                "summary": "Update Product",
                "description": "Update Product",
                "operationId": "update product",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Auth Product"
                ],
                "summary": "Delete product",
                "description": "Delete product",
                "operationId": "37bee50afc66585fe1f4d5fe77472b2c",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get list of products as a guest",
                "description": "Returns list of products",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserLoginRequest": {
                "title": "User Log In request",
                "description": "User Login request body data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email",
                        "type": "string",
                        "example": "test@bk.ru"
                    },
                    "password": {
                        "title": "password",
                        "description": "User password",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "UserRegistrationRequest": {
                "title": "User Registration request",
                "description": "User Registration request body data",
                "required": [
                    "email",
                    "name",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name",
                        "type": "string",
                        "example": "username"
                    },
                    "email": {
                        "title": "email",
                        "description": "Email",
                        "type": "string",
                        "example": "test@bk.ru"
                    },
                    "password": {
                        "title": "password",
                        "description": "User password",
                        "type": "string",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "User password_confirmation",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "StoreProductRequest": {
                "title": "Store product request",
                "description": "Store product request body data",
                "required": [
                    "name",
                    "description",
                    "category_id"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name",
                        "type": "string",
                        "example": "Product name"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description",
                        "type": "string",
                        "example": "Product description"
                    },
                    "category_id": {
                        "title": "category_id",
                        "description": "Category id",
                        "type": "integer",
                        "example": 1
                    },
                    "images[]": {
                        "title": "images",
                        "description": "Image",
                        "type": "array",
                        "items": {
                            "type": "file",
                            "format": "binary"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}